requirements.txt
matplotlib==3.1.3
opencv_python==4.2.0.32
torchvision==0.6.0a0+82fd1c8
numpy==1.18.1
scipy==1.4.1
scikit_image==0.16.2
torch==1.5.0
Pillow==7.2.0
skimage==0.0
